+++ core/civilization/person/base.py
@@ -3,12 +3,12 @@ from typing import Optional
 from pydantic import BaseModel
 
 from core.civilization.god.system import System
-from core.logging import ANSI, Color, Style, logger
+from core.logging import ANSI, Color, Style
 
-from .action import Action, ActionType
 from .brain import BaseBrain
 from .organize import BaseOrganize
 from .tool import BaseTool
+from .tracer import BasePersonTracer, PersonTracerWrapper
 
 
 class InviteParams(BaseModel):
@@ -68,6 +68,13 @@ class BasePerson(BaseModel, PersonMessageFormat):
     brain: BaseBrain = None
     friends: dict[str, "BasePerson"] = {}
     organize: BaseOrganize = None
+    tracer: PersonTracerWrapper = PersonTracerWrapper()
+
+    def set_tracers(self, tracers: list[BasePersonTracer]):
+        self.tracer = PersonTracerWrapper(tracers=tracers)
+
+    def add_tracer(self, tracer: BasePersonTracer):
+        self.tracer.add(tracer)
 
     def __str__(self):
         return ANSI((f"{self.name}({self.__class__.__name__})").center(20)).to(
@@ -76,129 +83,3 @@ class BasePerson(BaseModel, PersonMessageFormat):
 
     class Config:
         arbitrary_types_allowed = True
-
-
-class Log:
-    def to_idea(log_level: str):
-        def decorator(func):
-            def wrapper(self, prompt: str):
-                idea = func(self, prompt)
-                try:
-                    getattr(logger, log_level)(
-                        ANSI("[idea] ").to(Color.rgb(0xF6, 0xBA, 0x6F)) + idea
-                    )
-                except KeyError as e:
-                    logger.error("Failed to log to_idea: " + str(e))
-
-                return idea
-
-            return wrapper
-
-        return decorator
-
-    def think(log_level: str):
-        def decorator(func):
-            def wrapper(self, idea: str):
-                thought = func(self, idea)
-                try:
-                    getattr(logger, log_level)(
-                        ANSI("[thought] ").to(Color.rgb(0x6D, 0xA9, 0xE4)) + thought
-                    )
-                except KeyError as e:
-                    logger.error("Failed to log think: " + str(e))
-
-                return thought
-
-            return wrapper
-
-        return decorator
-
-    def to_actions(log_level: str):
-        def decorator(func):
-            def wrapper(self, thought: str):
-                actions = func(self, thought)
-                try:
-                    getattr(logger, log_level)(
-                        ANSI("[actions] ").to(Color.rgb(0xAD, 0xE4, 0xDB))
-                        + ", ".join([str(action.type) for action in actions])
-                    )
-                except KeyError as e:
-                    logger.error("Failed to log to_actions: " + str(e))
-
-                return actions
-
-            return wrapper
-
-        return decorator
-
-    def respond(log_level: str):
-        def decorator(func):
-            def wrapper(
-                self: "BasePerson",
-                sender: "BasePerson",
-                prompt: str,
-                params: TalkParams,
-            ):
-                result: str = func(self, sender, prompt, params)
-
-                try:
-                    getattr(logger, log_level)(
-                        str(self)
-                        + str(ActionType.Respond)
-                        + str(sender)
-                        + " | "
-                        + ANSI(result.split(System.PROMPT_SEPARATOR)[1].strip()).to(
-                            Color.white()
-                        )
-                    )
-                except KeyError as e:
-                    logger.error("Failed to log respond: " + str(e))
-
-                return result
-
-            return wrapper
-
-        return decorator
-
-    def act(log_level: str):
-        def get_target(self, type: ActionType, name: str):
-            if type in [ActionType.Invite, ActionType.Talk]:
-                return self.friends[name]
-            elif type in [ActionType.Build, ActionType.Use]:
-                return self.tools[name]
-
-        def decorator(func):
-            def wrapper(self, action: Action):
-                f = lambda target: (
-                    str(self)
-                    + str(action.type)
-                    + ANSI(str(target)).to(Color.green(), Style.bold())
-                    + " | "
-                    + ANSI(action.instruction).to(Color.white())
-                    + "\n"
-                    + ANSI(action.extra).to(Style.dim())
-                )
-
-                if action.type in [ActionType.Talk, ActionType.Use]:
-                    try:
-                        getattr(logger, log_level)(
-                            f(get_target(self, action.type, action.name))
-                        )
-                    except KeyError as e:
-                        logger.error("Failed to log act: " + str(e))
-
-                result = func(self, action)
-
-                if action.type in [ActionType.Invite, ActionType.Build]:
-                    try:
-                        getattr(logger, log_level)(
-                            f(get_target(self, action.type, action.name))
-                        )
-                    except KeyError as e:
-                        logger.error("Failed to log act: " + str(e))
-
-                return result
-
-            return wrapper
-
-        return decorator
